# artrack/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/artrack/auth/register": {
      "post": {
        "tags": [
          "artrack-auth"
        ],
        "summary": "Register",
        "description": "Register a new user",
        "operationId": "register_artrack_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artrack/auth/login": {
      "post": {
        "tags": [
          "artrack-auth"
        ],
        "summary": "Login",
        "description": "Login user",
        "operationId": "login_artrack_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artrack/auth/me": {
      "get": {
        "tags": [
          "artrack-auth"
        ],
        "summary": "Get Me",
        "description": "Get current user info",
        "operationId": "get_me_artrack_auth_me_get",
        "parameters": [
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artrack/auth/google": {
      "post": {
        "tags": [
          "artrack-auth"
        ],
        "summary": "Google Login",
        "description": "Login/Register via Firebase Google ID token and return api_key.",
        "operationId": "google_login_artrack_auth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artrack/auth/refresh": {
      "post": {
        "tags": [
          "artrack-auth"
        ],
        "summary": "Refresh Token",
        "description": "Refresh access token (placeholder - in real app would use refresh token)",
        "operationId": "refresh_token_artrack_auth_refresh_post",
        "parameters": [
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/artrack/sync/status": {
      "get": {
        "tags": [
          "artrack-sync"
        ],
        "summary": "Get Sync Status",
        "description": "Get sync status for current user",
        "operationId": "get_sync_status_artrack_sync_status_get",
        "parameters": [
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artrack/sync/retry-failed": {
      "post": {
        "tags": [
          "artrack-sync"
        ],
        "summary": "Retry Failed Uploads",
        "description": "Retry failed uploads",
        "operationId": "retry_failed_uploads_artrack_sync_retry_failed_post",
        "parameters": [
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artrack/sync/user-stats": {
      "get": {
        "tags": [
          "artrack-sync"
        ],
        "summary": "Get User Stats",
        "description": "Get detailed user statistics",
        "operationId": "get_user_stats_artrack_sync_user_stats_get",
        "parameters": [
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artrack/sync/cleanup": {
      "post": {
        "tags": [
          "artrack-sync"
        ],
        "summary": "Cleanup Failed Data",
        "description": "Clean up failed uploads and orphaned data",
        "operationId": "cleanup_failed_data_artrack_sync_cleanup_post",
        "parameters": [
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "trust_level": {
            "type": "string",
            "title": "Trust Level"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 86400
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "access_token",
          "api_key",
          "trust_level"
        ],
        "title": "AuthResponse"
      },
...
      "GoogleAuthRequest": {
        "properties": {
          "id_token": {
            "type": "string",
            "title": "Id Token"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "GoogleAuthRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuotaInfo": {
        "properties": {
          "storage_bytes": {
            "type": "integer",
            "title": "Storage Bytes"
          },
          "storage_limit": {
            "type": "integer",
            "title": "Storage Limit"
          },
          "uploads_this_month": {
            "type": "integer",
            "title": "Uploads This Month"
          },
          "upload_limit": {
            "type": "integer",
            "title": "Upload Limit"
          }
        },
        "type": "object",
        "required": [
          "storage_bytes",
          "storage_limit",
          "uploads_this_month",
          "upload_limit"
        ],
        "title": "QuotaInfo"
      },
...
      "SyncStatus": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "last_sync_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Sync At"
          },
          "pending_uploads": {
            "type": "integer",
            "title": "Pending Uploads"
          },
          "pending_analysis": {
            "type": "integer",
            "title": "Pending Analysis"
          },
          "failed_uploads": {
            "type": "integer",
            "title": "Failed Uploads"
          },
          "quota_used": {
            "$ref": "#/components/schemas/QuotaInfo"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "last_sync_at",
          "pending_uploads",
          "pending_analysis",
          "failed_uploads",
          "quota_used"
        ],
        "title": "SyncStatus"
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "display_name",
          "device_id"
        ],
        "title": "UserCreate"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "device_id"
        ],
        "title": "UserLogin"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "trust_level": {
            "type": "string",
            "title": "Trust Level"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "trust_level",
          "created_at"
        ],
        "title": "UserResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  }
}